FEAT[DEV]: Create base layout with navbar and sidebar
Created the initial structure of the application, including a navbar and sidebar. The layout follows a modular approach to facilitate future enhancements and styling. Added necessary libraries such as FontAwesome for icons and implemented a Context Provider to manage global state.

This commit is associated with the issue https://app.clickup.com/t/86c2vgvak