CHORE[DEV/PM]: Updated README, added design slides and organized repo
Updated the README with a translated version (EN), added information about the technologies used, and clarified the structure.
Uploaded the design slides used for the team presentation.
Minor cleanup of project files for better organization.

This commit is associated with the issue https://app.clickup.com/t/86c2vgw4m